Editor/Sidebar: Lift out a CategorySelector component
