+++ src/renderer/screens/Editor/Sidebar/ConsumerKeys.js
@@ -17,64 +17,25 @@
 
 import React from "react";
 
-import Button from "@material-ui/core/Button";
-import Dialog from "@material-ui/core/Dialog";
 import { withStyles } from "@material-ui/core/styles";
 
-import KeyButtonList from "../components/KeyButtonList";
-import Collapsible from "../components/Collapsible";
-import { KeymapDB } from "../../../../api/keymap";
-
-const db = new KeymapDB();
+import CategorySelector from "../components/CategorySelector";
 
 const styles = () => ({});
 
 class ConsumerKeysBase extends React.Component {
-  state = {
-    pickerOpen: false
-  };
-
-  openPicker = () => {
-    this.setState({ pickerOpen: true });
-  };
-
-  closePicker = () => {
-    this.setState({ pickerOpen: false });
-  };
-
-  onKeyChange = keyCode => {
-    this.props.onKeyChange(keyCode);
-    this.closePicker();
-  };
-
   render() {
-    const { keymap, selectedKey, layer } = this.props;
-
-    const key = keymap.custom[layer][selectedKey];
-    const label = db.format(key, "full");
+    const { keymap, selectedKey, layer, onKeyChange } = this.props;
 
     return (
-      <React.Fragment>
-        <Collapsible
-          expanded={db.isInCategory(key.code, "consumer")}
-          title="Consumer control keys"
-        >
-          <Button variant="contained" onClick={this.openPicker}>
-            {label.hint} {label.main}
-          </Button>
-        </Collapsible>
-        <Dialog
-          open={this.state.pickerOpen}
-          onClose={this.closePicker}
-          fullWidth
-          maxWidth="lg"
-        >
-          <KeyButtonList
-            keys={db.selectCategory("consumer")}
-            onKeyChange={this.onKeyChange}
-          />
-        </Dialog>
-      </React.Fragment>
+      <CategorySelector
+        title="Consumer control keys"
+        category="consumer"
+        keymap={keymap}
+        selectedKey={selectedKey}
+        layer={layer}
+        onKeyChange={onKeyChange}
+      />
     );
   }
 }
